+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/services/TradeService.java
@@ -1,8 +1,13 @@
 package com.nnk.springboot.services;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import javax.transaction.Transactional;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import com.nnk.springboot.domain.Trade;
@@ -10,23 +15,67 @@ import com.nnk.springboot.repositories.TradeRepository;
 
 @Service
 public class TradeService {
+  
+  private static Logger logger = LoggerFactory.getLogger(TradeService.class);
 
-  @Autowired
   private TradeRepository tradeRepository;
   
+  public TradeService(TradeRepository tradeRepository) {
+    this.tradeRepository = tradeRepository;
+  }
+
   public List<Trade> getAllTrade(){
+    logger.info("Getting a list of all Trade");
     return tradeRepository.findAll();
   }
   
+  @Transactional
   public Trade saveTrade(Trade trade) {
+    if(Objects.isNull(trade)) {
+      logger.error("New Trade must not be null"); 
+      return null;
+    }
+    logger.info("Creating a New Trade");
     return tradeRepository.save(trade);
   }
   
-  public Trade getTradeById(Integer id) {
-    return tradeRepository.findById(id).orElse(null);
+  @Transactional
+  public Trade updateTrade(Integer id, Trade trade) {
+    Optional<Trade> optTrade = tradeRepository.findById(id);
+    if(optTrade.isPresent()) {
+      Trade updatingTrade = optTrade.get();
+      updatingTrade.setAccount(trade.getAccount());
+      updatingTrade.setType(trade.getType());
+      updatingTrade.setBuyQuantity(trade.getBuyQuantity());
+      logger.info("Trade " + id + " updated");
+      return tradeRepository.save(updatingTrade);
+    } else {
+      logger.error("Trade with id: " + id + " not found nor updated");
+      return null;
+    }
   }
   
-  public void deleteById(Integer id) {
-    tradeRepository.deleteById(id);
+  public Trade getTradeById(Integer id) {
+    Optional<Trade> optTrade = tradeRepository.findById(id);
+    if(optTrade.isPresent()) {
+      logger.info("Trade with id " + id + " found");
+      return optTrade.get();
+    } else {
+      logger.error("Trade with id: " + id + " not found");
+      return null;
+    }
+  }
+  
+  @Transactional
+  public boolean deleteById(Integer id) {
+    Optional<Trade> optTrade = tradeRepository.findById(id);
+    if(optTrade.isPresent()) {
+      tradeRepository.deleteById(id);
+      logger.info("Trade with id " + id + " deleted");
+      return true;
+  } else {
+      logger.error("Trade with id: " + id + " not found nor deleted");
+      return false;
+  }
   }
 }
